// src/common/enums/order.enum.ts
export enum OrderStatusEnum {
    pending = 0,
    placed = 1,
    onway = 2,
    delivered = 3,
    canceled = 4
}



export enum OrderStatusNameEnum {
    pending = 'pending',
    placed = 'placed',
    onway = "onway",
    delivered = 'delivered',
    canceled = 'canceled'
}


export enum PaymentEnum {
    card = 'card',
    cash = 'cash',
}